#52: comparative word search in oXygen
For each of the terms we discussed, I conducted a search of the English translation XML files using oXygen. There is a separate sheet with a short excerpt and what folio it is found on. With the exception of  "as" and "as if," all of the searches to generate these sheets were NOT exact phrase searches, so there may be instances of variations of the term (for example, in the "like" sheet, we might see "likeness").