Fall 2021 final projects - terms related to red
These are search results from oxygen of terms related to red color in some way, where each csv contains a short excerpt and what folio it is found on. Naming scheme of the files: [semester]_[year]_[ms-version]_[exact-search-term].

Each csv contains the results from a search of either the tl (English translation) or the tcn (normalized French transcription) XML files using oxygen. In oxygen, used "enable XML search options" and chose only "element contents" to avoid getting results from element names or attributes (for example, "version" when searching for "ver"). These were also NOT exact phrase searches, in order to capture as many instances as possible, so there may be instances of variations of the term (for example, in the "red" sheet, we might see "redness"). For a couple of the term, we search with a leading white space (i.e., ` [term]`) - these are named with "space" before the term.